@@ -1,9 +1,9 @@
 package com.schoovello.cookerconnector.model;
 
-import android.arch.lifecycle.LiveData;
-import android.arch.lifecycle.MutableLiveData;
-import android.arch.lifecycle.ViewModel;
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.MutableLiveData;
+import androidx.lifecycle.ViewModel;
 
 import com.schoovello.cookerconnector.data.AlarmSpecLiveData;
 import com.schoovello.cookerconnector.datamodels.AlarmSpec;
