@@ -1,9 +1,10 @@
 package com.schoovello.cookerconnector.service;
 
 import android.app.Notification;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.NotificationManagerCompat;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.core.app.NotificationManagerCompat;
 
 import java.util.HashMap;
 import java.util.HashSet;
