@@ -1,7 +1,8 @@
 package com.schoovello.cookerconnector.sensorhub.activity;
 
 import android.os.Bundle;
-import android.support.v4.app.FragmentActivity;
+
+import androidx.fragment.app.FragmentActivity;
 
 import com.schoovello.cookerconnector.sensorhub.service.SensorHubService;
 
