@@ -1,6 +1,6 @@
 package com.schoovello.cookerconnector.sensorhub.registry;
 
-import android.arch.lifecycle.LiveData;
+import androidx.lifecycle.LiveData;
 
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
